Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
Django==6.0.2
django-redis==6.0.0
djangorestframework==3.16.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
msgpack==1.1.2
multidict==6.7.1
onetimepass==1.0.1
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyotp==2.9.0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sarvamai==0.1.23
six==1.17.0
sqlparse==0.5.5
twilio==9.10.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yarl==1.22.0
djangorestframework-simplejwt==5.4.0
